+++ src/main/java/com/soeun/GiftFunding/controller/ProductController.java
@@ -3,22 +3,18 @@ package com.soeun.GiftFunding.controller;
 import com.soeun.GiftFunding.dto.SaveProduct;
 import com.soeun.GiftFunding.dto.SearchProduct;
 import com.soeun.GiftFunding.dto.SearchProduct.Response;
-import com.soeun.GiftFunding.repository.ProductRepository;
-import com.soeun.GiftFunding.repository.elastic.ProductSearchRepository;
 import com.soeun.GiftFunding.service.ProductService;
-import java.util.List;
+import com.soeun.GiftFunding.service.ProductServiceImpl;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
-import org.elasticsearch.index.query.QueryBuilder;
-import org.elasticsearch.index.query.QueryBuilders;
 import org.springframework.data.domain.Page;
+import org.springframework.data.domain.PageRequest;
 import org.springframework.data.domain.Pageable;
-import org.springframework.data.elasticsearch.core.query.NativeSearchQuery;
-import org.springframework.data.elasticsearch.core.query.NativeSearchQueryBuilder;
+import org.springframework.data.domain.Sort;
 import org.springframework.http.ResponseEntity;
+import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
-import org.springframework.web.bind.annotation.RequestHeader;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 
@@ -33,7 +29,13 @@ public class ProductController {
         productService.saveProduct(saveProduct);
         return ResponseEntity.ok().build();
     }
+    @GetMapping
+    public ResponseEntity<Page<Response>> getProductList() {
 
+        return ResponseEntity.ok(
+            productService.getProductList()
+        );
+    }
     @PostMapping("/search")
     public ResponseEntity<Page<Response>> search(
         @RequestBody SearchProduct.Request request,
